fix: test islands
